+++ sgw-20180511/core/Models/DeleteElasticGatewayPrivateZoneResponse.cs
@@ -10,15 +10,15 @@ namespace AlibabaCloud.SDK.Sgw20180511.Models
 {
     public class DeleteElasticGatewayPrivateZoneResponse : TeaModel {
         [NameInMap("headers")]
-        [Validation(Required=true)]
+        [Validation(Required=false)]
         public Dictionary<string, string> Headers { get; set; }
 
         [NameInMap("statusCode")]
-        [Validation(Required=true)]
+        [Validation(Required=false)]
         public int? StatusCode { get; set; }
 
         [NameInMap("body")]
-        [Validation(Required=true)]
+        [Validation(Required=false)]
         public DeleteElasticGatewayPrivateZoneResponseBody Body { get; set; }
 
     }