+++ sgw-20180511/core/Models/DescribeGatewayImagesResponse.cs
@@ -10,15 +10,15 @@ namespace AlibabaCloud.SDK.Sgw20180511.Models
 {
     public class DescribeGatewayImagesResponse : TeaModel {
         [NameInMap("headers")]
-        [Validation(Required=true)]
+        [Validation(Required=false)]
         public Dictionary<string, string> Headers { get; set; }
 
         [NameInMap("statusCode")]
-        [Validation(Required=true)]
+        [Validation(Required=false)]
         public int? StatusCode { get; set; }
 
         [NameInMap("body")]
-        [Validation(Required=true)]
+        [Validation(Required=false)]
         public DescribeGatewayImagesResponseBody Body { get; set; }
 
     }