+++ sgw-20180511/core/Models/DescribeGatewayResponseBody.cs
@@ -95,6 +95,10 @@ namespace AlibabaCloud.SDK.Sgw20180511.Models
         [Validation(Required=false)]
         public string GatewayVersion { get; set; }
 
+        [NameInMap("HighAvailability")]
+        [Validation(Required=false)]
+        public bool? HighAvailability { get; set; }
+
         [NameInMap("InnerIp")]
         [Validation(Required=false)]
         public string InnerIp { get; set; }
@@ -166,6 +170,14 @@ namespace AlibabaCloud.SDK.Sgw20180511.Models
         [NameInMap("Type")]
         [Validation(Required=false)]
         public string Type { get; set; }
+
+        [NameInMap("UntrustedEnvInstanceType")]
+        [Validation(Required=false)]
+        public string UntrustedEnvInstanceType { get; set; }
+
+        [NameInMap("UntrustedEnvOssEndpoint")]
+        [Validation(Required=false)]
+        public string UntrustedEnvOssEndpoint { get; set; }
 
         [NameInMap("VSwitchId")]
         [Validation(Required=false)]