+++ sgw-20180511/ChangeLog.md
@@ -1,3 +1,12 @@
+2024-03-14 Version: 1.0.4
+- Update API CreateGateway: add param SecondaryVSwitchId.
+- Update API CreateGateway: add param UntrustedEnvId.
+- Update API CreateGateway: add param UntrustedEnvInstanceType.
+- Update API CreateGatewayCacheDisk: add param PerformanceLevel.
+- Update API DescribeGateway: update response param.
+- Update API DescribeGateways: update response param.
+
+
 2023-12-14 Version: 1.0.3
 - Generated csharp 2018-05-11 for sgw.